Fixed SSh bit

+++ components/ssh.sh
@@ -3,23 +3,40 @@
 # Has to be called after user creation
 #-------------------------------------------------
 
+sshd_config_file="/etc/ssh/sshd_config"
 
-# Quicker ssh login
-    echo 'UseDNS no' > /etc/ssh/sshd_config
+# Install SSH client and daemon
+    ${INSTALL} openssh
+
+
+# Enable SSH daemon
+    systemctl enable sshd.service
+# Quicker SSH login
+    echo 'UseDNS no' > ${sshd_config_file}
 # Enable SSHFS
-    echo "Subsystem sftp /usr/lib/sftp-server" >> /etc/ssh/sshd_config
+    echo "Subsystem sftp /usr/lib/sftp-server" >> ${sshd_config_file}
 
 # Security
     # Whitelist SSH port on the firewall
         ufw allow 22
     # Disable ssh connection to root account
-        echo "PermitRootLogin no" >> /etc/ssh/sshd_config
+        echo "PermitRootLogin no" >> ${sshd_config_file}
     # Disable authentication via password, require ssh key
-        echo "PasswordAuthentication no" >> /etc/ssh/sshd_config
+        echo "PasswordAuthentication no" >> ${sshd_config_file}
+
     # Add known_hosts file to user's ssh configuration
+    if [ -f "${AUTHORIZED_KEYS_FILE}" ]
+    then
         mkdir -p /home/${USERNAME}/.ssh/
         cp ${AUTHORIZED_KEYS_FILE} /home/${USERNAME}/.ssh/authorized_keys
         chown -R ${USERNAME}:${USERNAME} /home/${USERNAME}/.ssh
+    else
+        start_red_text="\e[00;31m"
+        stop_red_text="\e[00m"
+        echo "${start_red_text}"
+        echo "Authorized_keys file \"${AUTHORIZED_KEYS_FILE}\" not found. You won't be able to connect via SSH"
+        echo "${start_red_text}"
+    fi
